model config options
