+++ README.md
@@ -64,3 +64,4 @@ Contributions are welcome! Please open an issue or submit a pull request.
 ## License
 
 MIT License.
+
